Deep link data can be read when native link button is clicked and app is sideloaded
Deep link data can be read when native link button is clicked and app is sideloaded
