--- Branch Monster Factor/AppDelegateAdapter.swift
@@ -1,32 +1,0 @@
-import SwiftUI
-import SwiftData
-import BranchSDK
-
-/*
-    Using Adapter since it is a Swift UI app
-    Handles the Branch initialization logic
- */
-
-class AppDelegateAdapter: NSObject, UIApplicationDelegate {
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        
-        Branch.setUseTestBranchKey(true)
-        Branch.enableLogging()
-        Branch.getInstance().initSession(launchOptions: launchOptions) { (params, error) in
-            print(params as? [String: AnyObject] ?? [:])
-        }
-        Branch.getInstance().validateSDKIntegration()
-        
-        return true
-    }
-    
-    func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
-        Branch.getInstance().application(app, open: url, options: options)
-        return true
-    }
-    
-    func application(_ application: UIApplication, continue userActivity: NSUserActivity, restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void) -> Bool {
-        Branch.getInstance().continue(userActivity)
-        return true
-    }
-}
