+++ Branch Monster Factor/Branch_Monster_FactorApp.swift
@@ -5,8 +5,9 @@
 //  Created by Guru Prasadh on 05/11/25.
 //
 
-import SwiftUI
+import BranchSDK
 import Foundation
+import SwiftUI
 
 /*
     Entry point for the app
@@ -14,21 +15,24 @@ import Foundation
  */
 @main
 struct Branch_Monster_FactorApp: App {
-    @UIApplicationDelegateAdaptor(AppDelegateAdapter.self) var appDelegate
+    @UIApplicationDelegateAdaptor(AppDelegate.self) private var appDelegate
     @State private var progress: MonsterProgress
-    @AppStorage("persistentMonsterLevel") private var storedMonsterLevel: Int = 1
+    @AppStorage("persistentMonsterLevel") private var storedMonsterLevel: Int =
+        1
     @AppStorage("persistentMonsterExp") private var storedMonsterExp: Double = 0
-    @AppStorage("persistentMonsterColor") private var storedMonsterColor: String = "yellow"
+    @AppStorage("persistentMonsterColor") private var storedMonsterColor:
+        String = "yellow"
 
     init() {
         let defaults = UserDefaults.standard
-        
+
         let initialLevel = defaults.integer(forKey: "persistentMonsterLevel")
         let initialExp = defaults.double(forKey: "persistentMonsterExp")
-        let initialColor = defaults.string(forKey: "persistentMonsterColor") ?? "yellow"
+        let initialColor =
+            defaults.string(forKey: "persistentMonsterColor") ?? "yellow"
 
         let safeLevel = initialLevel == 0 ? 1 : initialLevel
-        
+
         _progress = State(
             initialValue: MonsterProgress(
                 initialXP: initialExp,
@@ -44,9 +48,42 @@ struct Branch_Monster_FactorApp: App {
                 OnboardingScreen()
             }
             .environment(progress)
-            .onChange(of: progress.monsterLevel) { storedMonsterLevel = progress.monsterLevel }
-            .onChange(of: progress.currentXP) { storedMonsterExp = progress.currentXP }
-            .onChange(of: progress.selectedColor) { storedMonsterColor = progress.selectedColor }
+            .onChange(of: progress.monsterLevel) {
+                storedMonsterLevel = progress.monsterLevel
+            }
+            .onChange(of: progress.currentXP) {
+                storedMonsterExp = progress.currentXP
+            }
+            .onChange(of: progress.selectedColor) {
+                storedMonsterColor = progress.selectedColor
+            }
         }
+    }
+}
+
+class AppDelegate: NSObject, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        
+        Branch.setUseTestBranchKey(true)
+        Branch.enableLogging()
+        Branch.getInstance().initSession(launchOptions: launchOptions) { (params, error) in
+            print(params as? [String: AnyObject] ?? [:])
+        }
+        Branch.getInstance().validateSDKIntegration()
+        
+        return true
+    }
+    
+    func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
+        print("Debugging - open URL hit")
+        Branch.getInstance().application(app, open: url, options: options)
+        return true
+    }
+    
+    func application(_ application: UIApplication, continue userActivity: NSUserActivity, restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void) -> Bool {
+        print("Debugging - continue user activity hit")
+        Branch.getInstance().continue(userActivity)
+        return true
     }
 }
